+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Fix language fall through logic
 
 ### v0.33.6
 * Consolidate language detection stuff in `language_helper.ts`